+++ familyhelper-clannad-stack/src/main/java/com/dwarfeng/familyhelper/clannad/stack/handler/PushHandler.java
@@ -1,6 +1,7 @@
 package com.dwarfeng.familyhelper.clannad.stack.handler;
 
 import com.dwarfeng.familyhelper.clannad.stack.bean.dto.BirthdayBlessInfo;
+import com.dwarfeng.familyhelper.clannad.stack.bean.entity.Message;
 import com.dwarfeng.subgrade.stack.exception.HandlerException;
 import com.dwarfeng.subgrade.stack.handler.Handler;
 
@@ -29,4 +30,13 @@ public interface PushHandler extends Handler {
      * @throws HandlerException 处理器异常。
      */
     void birthdayBlessHappened(List<BirthdayBlessInfo> birthdayBlessInfos) throws HandlerException;
+
+    /**
+     * 留言发送时执行的调度。
+     *
+     * @param message 留言。
+     * @throws HandlerException 处理器异常。
+     * @since 1.5.0
+     */
+    void messageSent(Message message) throws HandlerException;
 }